Fixed the stripe payment

+++ database/migrations/2022_05_16_015417_create_carts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users');
             $table->boolean('checkout')->default(false);
-            $table->bigInteger('total_price');
+            $table->float('total_price');
             $table->timestamps();
             $table->softDeletes();
         });
